Minor registration cleanup for Windows
Clean up parts of the Windows registration to use the correct terms for
a couple of things, and change the App Registration to use the EXE name
and not the canonical name (it's supposed to be the EXE name, and they
used to match, but they don't any more).

This should make the code more closely match the upcoming blog post.