Bump some version numbers to invalidate cache
My previous change to work around the GH workflow cache action bug on
macOS didn't actually work because I happened to restore Cargo.lock to
the exact version I had earlier, so it got a direct match on the cache.

Upgrade a few minor packages to get a new hash.